refactor: move JitEvm to crates/revmc/src/revm_evm.rs
Move the JitEvm EvmTr wrapper from revmc-cli into the revmc crate so it
can be shared across crates. Re-exported as revmc::JitEvm.
