+++ crates/revmc/Cargo.toml
@@ -29,6 +29,7 @@ revmc-llvm = { workspace = true, optional = true }
 alloy-primitives = { workspace = true, features = ["std", "map-fxhash"] }
 revm-bytecode.workspace = true
 revm-context-interface.workspace = true
+revm-handler.workspace = true
 revm-interpreter.workspace = true
 revm-primitives.workspace = true
 revm-state.workspace = true
