+++ crates/revmc/src/lib.rs
@@ -15,6 +15,9 @@ pub use compiler::{EvmCompiler, EvmCompilerInput};
 mod linker;
 pub use linker::{Linker, shared_library_path};
 
+mod revm_evm;
+pub use revm_evm::JitEvm;
+
 /// Internal tests and testing utilities. Not public API.
 #[cfg(any(test, feature = "__fuzzing"))]
 pub mod tests;
